Extract fix helpers
